Catch all exceptions In async functions.
Prevents crashes in case of failure like
https://github.com/googlesamples/android-play-billing/issues/26.

All exceptions catched in async threads will now be reported in console
and then retruned to listener as `BILLING_RESPONSE_RESULT_ERROR`. If
SecurityException will be raised - which can happen if billing service
manifest changes - `BILLING_RESPONSE_RESULT_BILLING_UNAVAILABLE` will
be returned. Although creation of the same error response in each case
like this might have performace implications, this code is clearly not in
the hot path and is unlikely to be called frequently. In about 5 years of
Google Play billing v3.0 existance this code was not usefull untill last
month and hopefully won't be usefull in the future. I would rather create
a new object than investigate any side effects that might emerge.